feat(Funcionarios): refactor employee data structure and import model in components
