+++ src/components/funcionarios/listarfuncionarios/ListarFuncionarios.tsx
@@ -1,19 +1,5 @@
+import { funcionarios } from "../../../models/Funcionarios";
 import CardFuncionario from "../cardfuncionario/CardFuncionario";
-
-const funcionarios = [
-    {
-        nome: "João",
-        cargo: "Desenvolvedor Fullstack",
-        foto: "https://bootdey.com/img/Content/avatar/avatar7.png",
-        departamento: "Tecnologia",
-    },
-    {
-        nome: "Maria",
-        cargo: "Desenvolvedor Backend",
-        foto: "https://bootdey.com/img/Content/avatar/avatar8.png",
-        departamento: "Tecnologia",
-    },
-];
 
 const ListarFuncionarios = () => {
     return (
